requirements.txt
aiohappyeyeballs @ file:///home/conda/feedstock_root/build_artifacts/aiohappyeyeballs_1741775197943/work
aiohttp @ file:///Users/runner/miniforge3/conda-bld/aiohttp_1749923443784/work
aiosignal @ file:///home/conda/feedstock_root/build_artifacts/aiosignal_1751626463503/work
annotated-types==0.7.0
anyio==4.9.0
async-timeout @ file:///home/conda/feedstock_root/build_artifacts/async-timeout_1733235340728/work
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1741918516150/work
Brotli @ file:///Users/runner/miniforge3/conda-bld/brotli-split_1749229914644/work
cachetools==5.5.2
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1749972191589/work/certifi
cffi @ file:///Users/runner/miniforge3/conda-bld/cffi_1725560622100/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1746214863626/work
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
dataclasses @ file:///home/conda/feedstock_root/build_artifacts/dataclasses_1628958434797/work
datasets @ file:///home/conda/feedstock_root/build_artifacts/datasets_1691593734461/work
dill @ file:///home/conda/feedstock_root/build_artifacts/dill_1690101045195/work
exceptiongroup==1.3.0
feedparser @ file:///home/conda/feedstock_root/build_artifacts/feedparser_1734808041952/work
filelock @ file:///home/conda/feedstock_root/build_artifacts/filelock_1741969488311/work
frozenlist @ file:///Users/runner/miniforge3/conda-bld/frozenlist_1746635428890/work
fsspec @ file:///home/conda/feedstock_root/build_artifacts/fsspec_1748101524683/work
gmpy2 @ file:///Users/runner/miniforge3/conda-bld/gmpy2_1745509426225/work
google-auth==2.40.3
google-genai==1.7.0
h11==0.16.0
h2 @ file:///home/conda/feedstock_root/build_artifacts/h2_1738578511449/work
hf-xet @ file:///Users/runner/miniforge3/conda-bld/bld/rattler-build_hf-xet_1750541254/work
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
httpcore==1.0.9
httpx==0.28.1
huggingface_hub @ file:///home/conda/feedstock_root/build_artifacts/huggingface_hub_1750865802145/work
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1733211830134/work
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_importlib-metadata_1747934053/work
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1741263328855/work
joblib @ file:///home/conda/feedstock_root/build_artifacts/joblib_1748019130050/work
lxml @ file:///Users/runner/miniforge3/conda-bld/lxml_1751021689128/work
markdown2 @ file:///home/conda/feedstock_root/build_artifacts/markdown2_1737823932681/work
MarkupSafe @ file:///Users/runner/miniforge3/conda-bld/markupsafe_1733219663979/work
mpmath @ file:///home/conda/feedstock_root/build_artifacts/mpmath_1733302684489/work
multidict @ file:///Users/runner/miniforge3/conda-bld/multidict_1751310641889/work
multiprocess @ file:///Users/runner/miniforge3/conda-bld/multiprocess_1695458927554/work
networkx @ file:///home/conda/feedstock_root/build_artifacts/networkx_1698504735452/work
numpy @ file:///Users/runner/miniforge3/conda-bld/numpy_1732314388626/work/dist/numpy-2.0.2-cp39-cp39-macosx_11_0_arm64.whl#sha256=f5ae29dd6f735c6dcad61765cffcfb32661c04e515c759b21399c7a3a39430a4
optree @ file:///Users/runner/miniforge3/conda-bld/optree_1748442503493/work
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1745345660/work
pandas @ file:///Users/runner/miniforge3/conda-bld/pandas_1749099750763/work
pillow @ file:///Users/runner/miniforge3/conda-bld/pillow_1751482050926/work
propcache @ file:///Users/runner/miniforge3/conda-bld/propcache_1744525055131/work
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybind11 @ file:///D:/bld/pybind11-split_1747934313674/work
pybind11_global @ file:///home/conda/feedstock_root/build_artifacts/pybind11-split_1747934270866/work
pycparser @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pycparser_1733195786/work
pydantic==2.11.7
pydantic_core==2.33.2
PyPDF2 @ file:///home/conda/feedstock_root/build_artifacts/pypdf2_1735229443413/work
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1733217236728/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_python-dateutil_1751104122/work
python-dotenv==1.1.1
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1742920838005/work
PyYAML @ file:///Users/runner/miniforge3/conda-bld/pyyaml_1737454732336/work
regex @ file:///Users/runner/miniforge3/conda-bld/regex_1730952212651/work
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1749498106507/work
rsa==4.9.1
safetensors @ file:///Users/runner/miniforge3/conda-bld/safetensors_1740651535429/work
scikit-learn @ file:///Users/runner/miniforge3/conda-bld/scikit-learn_1736496791599/work/dist/scikit_learn-1.6.1-cp39-cp39-macosx_11_0_arm64.whl#sha256=aaca544017450d0f614af6e4c06601e2aa132d6dfbb957b49e7c5fc3b37f6417
scipy @ file:///Users/runner/miniforge3/conda-bld/scipy-split_1716470312029/work/dist/scipy-1.13.1-cp39-cp39-macosx_11_0_arm64.whl#sha256=95cc9a6aceed50037865da4a45baac05fd0dcee7e557c6149fde270f2ffebafc
sentence-transformers @ file:///home/conda/feedstock_root/build_artifacts/sentence-transformers_1751402683678/work
sgmllib3k @ file:///home/conda/feedstock_root/build_artifacts/sgmllib3k_1734543040599/work
six @ file:///home/conda/feedstock_root/build_artifacts/six_1733380938961/work
sniffio==1.3.1
sympy @ file:///home/conda/feedstock_root/build_artifacts/sympy_1745946051654/work
threadpoolctl @ file:///home/conda/feedstock_root/build_artifacts/threadpoolctl_1741878222898/work
tokenizers @ file:///Users/runner/miniforge3/conda-bld/tokenizers_1751645861052/work/bindings/python
torch @ file:///Users/runner/miniforge3/conda-bld/libtorch_1751416286009/work
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1735661334605/work
transformers @ file:///home/conda/feedstock_root/build_artifacts/transformers_1751628914573/work
typing-inspection==0.4.1
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1751643513/work
tzdata @ file:///home/conda/feedstock_root/build_artifacts/python-tzdata_1742745135198/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1750271362675/work
websockets==15.0.1
xxhash @ file:///Users/runner/miniforge3/conda-bld/python-xxhash_1740594853298/work
yarl @ file:///Users/runner/miniforge3/conda-bld/yarl_1749554901307/work
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1749421620841/work
zstandard==0.23.0
